project/transaction_service/requirements.txt
flask
kafka-python
redis
flask_cors